Adds simple list function and test data
This adds a listBoardGames method that will print out a readable summary
about the board games available to play.

Also adds test data with two board games, in JSON format.

I decided to use JSON as I am familiar with writing it and parsing it using
other languages, and because it is simple and easy to read.